+++ Apps/WebAPI/sdLitica.Management.WebAPI/Models/JsonAPI/PaginationProperties.cs
