+++ Apps/WebAPI/sdLitica.Management.WebAPI/Models/Management/TokenJsonEntity.cs
