+++ Apps/WebAPI/sdLitica.Management.WebAPI/Models/Management/UserJsonEntity.cs
