+++ Apps/WebAPI/sdLitica.Management.WebAPI/Models/Security/CustomAuthHandler.cs
@@ -2,7 +2,6 @@
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Net.Http.Headers;
-using sdLitica.Services.Management;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,6 +10,7 @@ using System.Security.Principal;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
+using sdLitica.PlatformCore;
 
 namespace sdLitica.WebAPI.Security
 {
