+++ Apps/WebAPI/sdLitica.Management.WebAPI/Models/Security/UserPrincipal.cs
