+++ Apps/WebAPI/sdLitica.Management.WebAPI/Properties/launchSettings.json
