+++ Apps/WebAPI/sdLitica.Management.WebAPI/sdLitica.Management.WebAPI.csproj
@@ -15,9 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\Entities\sdLitica.Entities\sdLitica.Entities.csproj" />
-    <ProjectReference Include="..\..\..\Infrastructure\sdLitica.Bootstrap\sdLitica.Bootstrap.csproj" />
-    <ProjectReference Include="..\..\..\Services\sdLitica.Services\sdLitica.Services.csproj" />
+    <ProjectReference Include="..\..\..\Infrastructure\Bootstrap\sdLitica.Management.Bootstrap\sdLitica.Management.Bootstrap.csproj" />
   </ItemGroup>
 
 </Project>
