+++ Data/sdLitica.Relational/sdLitica.Relational.csproj
@@ -9,7 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Entities\sdLitica.Entities\sdLitica.Entities.csproj" />
+    <ProjectReference Include="..\..\Domain\sdLitica.Entities\sdLitica.Entities.csproj" />
   </ItemGroup>
 
 </Project>
