+++ Domain/sdLitica.Entities.Test/sdLitica.Entities.Test.csproj
