+++ Domain/sdLitica.Entities/Abstractions/Entity.cs
