+++ Domain/sdLitica.Entities/Abstractions/IEntity.cs
