+++ Domain/sdLitica.Entities/Abstractions/IRepositoryBase.cs
