+++ Domain/sdLitica.Entities/Management/Repositories/IUserRepository.cs
