+++ Domain/sdLitica.Entities/Management/User.cs
