+++ Domain/sdLitica.Entities/Management/UserToken.cs
