+++ Domain/sdLitica.Entities/sdLitica.Entities.csproj
