+++ Infrastructure/Bootstrap/sdLitica.Management.Bootstrap/Settings/BootstrapSettings.cs
