+++ Infrastructure/Bootstrap/sdLitica.Management.Bootstrap/sdLitica.Management.Bootstrap.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\Data\sdLitica.Relational\sdLitica.Relational.csproj" />
+    <ProjectReference Include="..\..\..\Domain\sdLitica.Entities\sdLitica.Entities.csproj" />
+    <ProjectReference Include="..\..\..\Services\sdLitica.PlatformCore\sdLitica.PlatformCore.csproj" />
+  </ItemGroup>
+
+</Project>
