--- Infrastructure/sdLitica.Bootstrap/sdLitica.Bootstrap.csproj
@@ -1,13 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\Entities\sdLitica.Entities\sdLitica.Entities.csproj" />
-    <ProjectReference Include="..\..\Persistance\sdLitica.Relational\sdLitica.Relational.csproj" />
-    <ProjectReference Include="..\..\Services\sdLitica.Services\sdLitica.Services.csproj" />
-  </ItemGroup>
-
-</Project>
