+++ Services/sdLitica.PlatformCore.Test/sdLitica.PlatformCore.Test.csproj
