+++ Services/sdLitica.PlatformCore/sdLitica.PlatformCore.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Entities\sdLitica.Entities\sdLitica.Entities.csproj" />
+    <ProjectReference Include="..\..\Domain\sdLitica.Entities\sdLitica.Entities.csproj" />
     <ProjectReference Include="..\..\Infrastructure\sdLitica.Exceptions\sdLitica.Exceptions.csproj" />
   </ItemGroup>
